Fix user inputs incorrectly affecting scores
Case 1: User input having named keys

It's not uncommon to want to take things like a user's name, username,
email address, and other user metadata into account when evaluating
password strength. These items may already already be available in a
string-keyed array such as `['name' => 'bob', 'email' =>
'bob@example.com']`. The present implementation will use the string key
as the input rank, which causes an error when attempting to take the
`log()` of it.

This is fixed by setting the rank of any user input that has a string
key to whatever its numerical position in the user input array is.

Case 1: Matching user input with rank 0 scoring incorrectly

`correct horse battery staple` with no user input would score a 4, but
the same password with `['c']` as the user input changes the score to 0
(which is obviously not correct). This is because a rank of 0 causes an
entropy score of `-INF`, which causes the crack time to evaluate to 0.

This is fixed by ensuring the minimum rank that can be assigned to a
value is `1`. Any reasonable number would work, but `1` seemed like a
fair default, since the default behavior is to use the array index, and
`1` is the lowest nonzero positive integer.